feat: ablation-calibrated functional scores + attention paradigm classification
- models.py: add attention_paradigm_of() — classifies model as mha/gqa/alternating
- caz.py: add functional_caz_score per region; ABLATION_PRIORS maps paradigm →
  fisher_trust multiplier (mha=1.0, gqa=0.35, alternating=0.1/boost at global layer)
- caz_registry.py: classify_caz() updated to use functional score; adds 'functional'
  and 'artifact' categories for architecture-calibrated classification
- dark_feature_scanner.py: fix dtype= kwarg (was torch_dtype=)

Required by cka_validation.py import of attention_paradigm_of.

+++ rosetta_tools/dark_feature_scanner.py
@@ -219,9 +219,9 @@ def scan_dark_features(
     log.info("Loading %s...", model_id)
     tokenizer = AutoTokenizer.from_pretrained(model_id)
     try:
-        model = AutoModel.from_pretrained(model_id, torch_dtype=dtype, device_map=device)
+        model = AutoModel.from_pretrained(model_id, dtype=dtype, device_map=device)
     except (ValueError, ImportError):
-        model = AutoModel.from_pretrained(model_id, torch_dtype=dtype).to(device)
+        model = AutoModel.from_pretrained(model_id, dtype=dtype).to(device)
     model.eval()
     log_vram("after model load")
